chore(metadata): reorganize pipeline, add doc
