+++ src/cli.rs
@@ -1,3 +1,4 @@
+//! Command line arguments and parameters management/parsing.
 use std::path::PathBuf;
 
 use structopt::StructOpt;
@@ -14,8 +15,25 @@ pub enum Ungoliant {
 
 #[derive(Debug, StructOpt)]
 /// Download command and parameters.
-// should it be merged with the "real" download one?
-// or at least download::Downloader implement From<Download>
+/// ```sh
+/// ungoliant-download 0.1.0
+/// Downloading of CommonCrawl
+///
+/// USAGE:
+///     ungoliant download [OPTIONS] <paths-file> <dst>
+///
+/// FLAGS:
+///     -h, --help       Prints help information
+///     -V, --version    Prints version information
+///
+/// OPTIONS:
+///     -t <n-tasks>        number of tokio tasks. Default is 4.
+///     -o <offset>         number of files to skip. Default is 0.
+///
+/// ARGS:
+///     <paths-file>    path to wet.paths file
+///     <dst>           download destination
+/// ```
 pub struct Download {
     #[structopt(parse(from_os_str), help = "path to wet.paths file")]
     pub paths_file: PathBuf,
@@ -28,6 +46,24 @@ pub struct Download {
 }
 
 #[derive(Debug, StructOpt)]
+/// Pipeline command and parameters.
+///
+/// ```sh
+/// ungoliant-pipeline 0.1.0
+/// Run pipeline
+///
+/// USAGE:
+///     ungoliant pipeline [FLAGS] <src> <dst>
+///
+/// FLAGS:
+///     -h, --help             Prints help information
+///     -V, --version          Prints version information
+///     -m, --with_metadata    extract metadata
+///
+/// ARGS:
+///     <src>    source (contains n.txt.gz)
+///     <dst>    pipeline result destination
+/// ```
 pub struct Pipeline {
     #[structopt(parse(from_os_str), help = "source (contains n.txt.gz)")]
     pub src: PathBuf,
