+++ src/pipeline/oscar_metadata/mod.rs
@@ -1,0 +1,5 @@
+mod chunks;
+mod metadata;
+pub mod oscar_metadata;
+
+use metadata::Metadata;
